ActivityStream refactoring : Moving packages of certain classes, updating samples and other consumers
