+++ samples/j2ee/com.ibm.sbt.sample.web/WebContent/samples/java/Social/ActivityStreams/Post Event Using Populator.jsp
@@ -17,10 +17,10 @@
 
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
 +<%@page import="com.ibm.commons.runtime.Context"%> 
-<%@page import="com.ibm.sbt.services.client.connections.activitystreams.templates.ASDataPopulator"%>
+<%@page import="com.ibm.sbt.services.client.connections.activitystreams.ASDataPopulator"%>
 <%@page import="java.io.PrintWriter"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.model.Reply"%>
-<%@page import="com.ibm.sbt.services.client.connections.activitystreams.model.ActivityStreamEntity"%>
+<%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamEntity"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamEntityList"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamService"%>
 <%@page import="java.util.List"%>
