+++ samples/j2ee/com.ibm.sbt.sample.web/WebContent/samples/java/Social/ActivityStreams/UpdatesFromUser.jsp
@@ -19,7 +19,7 @@
 <%@page import="com.ibm.commons.runtime.Context"%>
 <%@page import="java.io.PrintWriter"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.model.Reply"%>
-<%@page import="com.ibm.sbt.services.client.connections.activitystreams.model.ActivityStreamEntity"%>
+<%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamEntity"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamEntityList"%>
 <%@page import="com.ibm.sbt.services.client.connections.activitystreams.ActivityStreamService"%>
 <%@page import="java.util.List"%>
