+++ src/eclipse/plugins/com.ibm.sbt.core/src/com/ibm/sbt/services/client/connections/activitystreams/ASDataPopulator.java
@@ -1,4 +1,4 @@
-package com.ibm.sbt.services.client.connections.activitystreams.templates;
+package com.ibm.sbt.services.client.connections.activitystreams;
 /*
  * © Copyright IBM Corp. 2013
  *
